+++ src/boot/DeviceManager.ts
@@ -66,7 +66,6 @@ export interface IDeviceInfo {
 
 ErrorManager.register(
     'DeviceManager',
-    'S5dBTBKTnVbF',
     'DM_DEVICE_NOT_FOUND',
     'Device not found', {
     device: Rule.string().required().example('Master').description('Device name')
@@ -74,7 +73,6 @@ ErrorManager.register(
 
 ErrorManager.register(
     'DeviceManager',
-    '2CJSDE95V9O1',
     'DM_LIST_NOT_FOUND',
     'list.json not found', {
     vendor: Rule.string().required().example('vrack').description('Device group name')
@@ -82,7 +80,6 @@ ErrorManager.register(
 
 ErrorManager.register(
     'DeviceManager',
-    'BPX96F93ZCJA',
     'DM_LIST_INCORRECT',
     'list.json is not correct', {
     vendor: Rule.string().required().example('vrack').description('Device group name')
@@ -90,13 +87,12 @@ ErrorManager.register(
 
 ErrorManager.register(
     'DeviceManager',
-    'H1NDNVCOLWST',
     'DM_VENDOR_NOT_FOUND',
     'Vendor not found', {
     vendor: Rule.string().required().example('vrack').description('Vendor name')
 })
 
-ErrorManager.register('DeviceManager', 'HQ62LW1YLTDE', 'DM_GET_INFO_EXCEPTION', 'Error retrieving device data, see error in attachment', {})
+ErrorManager.register('DeviceManager', 'DM_GET_INFO_EXCEPTION', 'Error retrieving device data, see error in attachment', {})
 
 
 /**