+++ src/validator/types/BooleanType.ts
@@ -46,6 +46,6 @@ export default class BooleanType  extends BasicType {
 }
 
 ErrorManager.register(
-    'Validator', 'Fr7BvAlZyZPm', 'VR_IS_NOT_BOOLEAN',
+    'Validator', 'VR_IS_NOT_BOOLEAN',
     'Value must be a boolean', {
 })